+++ libs/web-ui-kit/src/composites/section/index.ts
@@ -1,2 +1,3 @@
+export * from './CollapsibleSection';
 export * from './GroupLabel';
 export * from './SectionHeader';